+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipeline.LightLoop.cs
@@ -405,6 +405,8 @@ namespace UnityEngine.Rendering.HighDefinition
                                     HDCamera            hdCamera,
                                     ref PrepassOutput   prepassOutput,
                                     TextureHandle       clearCoatMask,
+                                    TextureHandle       rayCountTexture,
+                                    Texture             skyTexture,
                                     bool                transparent)
         {
             if (!hdCamera.IsSSREnabled(transparent))
@@ -412,14 +414,15 @@ namespace UnityEngine.Rendering.HighDefinition
 
             TextureHandle result;
 
-            // TODO RENDERGRAPH
-            //var settings = hdCamera.volumeStack.GetComponent<ScreenSpaceReflection>();
-            //bool usesRaytracedReflections = hdCamera.frameSettings.IsEnabled(FrameSettingsField.RayTracing) && settings.rayTracing.value;
-            //if (usesRaytracedReflections)
-            //{
-            //    RenderRayTracedReflections(hdCamera, cmd, m_SsrLightingTexture, renderContext, m_FrameCount, true);
-            //}
-            //else
+            var settings = hdCamera.volumeStack.GetComponent<ScreenSpaceReflection>();
+            bool usesRaytracedReflections = hdCamera.frameSettings.IsEnabled(FrameSettingsField.RayTracing) && settings.rayTracing.value;
+            if (usesRaytracedReflections)
+            {
+                result = RenderRayTracedReflections(renderGraph, hdCamera,
+                                                    prepassOutput.depthBuffer, prepassOutput.stencilBuffer, prepassOutput.normalBuffer, prepassOutput.resolvedMotionVectorsBuffer, clearCoatMask, skyTexture, rayCountTexture,
+                                                    m_FrameCount, m_ShaderVariablesRayTracingCB, transparent);
+            }
+            else
             {
                 if (transparent)
                     BuildCoarseStencilAndResolveIfNeeded(renderGraph, hdCamera, ref prepassOutput);